Update Objects and Classes in C++ 2
Ex2: A class for rational numbers, the number should always be represented as a
fully reduced fraction of the form
𝑛𝑢𝑚𝑒𝑟𝑎𝑡𝑜𝑟
𝑑𝑒𝑛𝑢𝑚𝑒𝑟𝑎𝑡𝑜𝑟
where denumerator>0
a. Write a class Rational, which represents a rational number. The constructor
should have the numerator and the denominator as arguments. Be sure to check
for valid input. In addition, the class has two functions numerator() and
denominator() that return the values of the numerator and denominator. The
class should have three constructors:
• a default constructor that initializes the fraction with 1,
• a constructor that initializes the fraction with a given numerator and
denominator and
• a constructor that initializes the fraction with a given whole number.
b. Writing a function int gcd(int,int) (greatest common divisor), you will need it
to reduce fractions.